# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental API",
        "description": "Dokumentacja systemu wypożyczalni",
        "version": "1.0.0"
    },
    "paths": {
...
        "/admin/promo-codes": {
            "get": {
                "tags": [
                    "Promo Codes"
                ],
                "summary": "Lista wszystkich kodów promocyjnych",
                "operationId": "getAdminPromoCodes",
                "responses": {
                    "200": {
                        "description": "Paginowana lista kodów wraz z relacjami",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Promo Codes"
                ],
                "summary": "Generowanie nowego kodu dla użytkownika",
                "operationId": "storeAdminPromoCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "amount"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID użytkownika, do którego trafi kod",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "maximum": 10000,
                                        "minimum": 1,
                                        "example": 50
                                    },
                                    "expires_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-12-31 23:59:59",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Bonus świąteczny",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kod utworzony pomyślnie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/promo-codes/{id}": {
            "delete": {
                "tags": [
                    "Promo Codes"
                ],
                "summary": "Usuwanie kodu promocyjnego",
                "operationId": "deleteAdminPromoCode",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID kodu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kod usunięty"
                    },
                    "422": {
                        "description": "Nie można usunąć użytego kodu"
                    },
                    "404": {
                        "description": "Kod nie istnieje"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}